+++ display_on_matrix/text/timed.py
@@ -1,21 +1,24 @@
 import asyncio
+import time
 from rgbmatrix import graphics
 
-async def ticker_text(self, offscreen_canvas, font, pos_x, pos_y, textcolor, text, loop, loops, speed):
+async def timed_text(self, offscreen_canvas, font, pos_x, pos_y, textcolor, text, loop, show_time, speed):
     try:
-        while loops > 0:
+        time_start = time.time()
+        while True :  #To interrupt after time is done or to finish run?
+            done_flag = False
             offscreen_canvas.Clear()
             length = graphics.DrawText(offscreen_canvas, font, pos_x, pos_y, textcolor, text)
             pos_x -= 1
 
             if (pos_x + length < 0):
-                loops -= 1
                 pos_x = offscreen_canvas.width
+                if time.time() > time_start + show_time: break
 
             offscreen_canvas = await loop.run_in_executor(None, self.matrix.SwapOnVSync, offscreen_canvas)
+            
             # adapt to modify speed
             await asyncio.sleep(speed)
-
     except asyncio.CancelledError:
-        print("cancel")
         return "error"
+    return
