+++ event_source/generate_source.py
@@ -5,13 +5,12 @@ from tests.generate_message import generate_random_event
 class RandomEventSource(EventSource):
     async def get_events(self):
         """Generate a list of random events."""
-        events = [generate_random_event() for _ in range(10)]  # Generate 10 random events
-        print(f"Generated events: {events}")#######
+        events = [generate_random_event() for _ in range(20)]  # Generate (int) random events
+        print(f"Generated events: {events}")
         return events
 
     async def write_event(self, event):
         """Continuously read from stdin and notify consumers."""
-        print(f"Writing event: {event}")#######
         await self._notify_consumers({"event": event})
 
     async def start(self):
@@ -20,7 +19,7 @@ class RandomEventSource(EventSource):
         events = await self.get_events()
         for event in events:
             await self.write_event(event)
-            await asyncio.sleep(5)
+            #await asyncio.sleep(5)
         await asyncio.sleep(120)
 
     async def stop(self):
